requirements.txt
streamlit==1.32.2
pandas==2.2.0
yfinance==0.2.36
plotly==5.19.0
PyYAML==6.0.1
fredapi==0.5.1
requests==2.31.0